fix: show URLs only for failed runs, +refactor
