+++ cmd/current.go
@@ -9,7 +9,6 @@ import (
 
 	ghapi "github.com/cli/go-gh/v2/pkg/api"
 	"github.com/dhth/act3/internal/gh"
-	"github.com/dhth/act3/ui"
 	"github.com/go-git/go-git/v5"
 )
 
@@ -21,15 +20,15 @@ var (
 	errCouldntParseRemoteURL = errors.New("couldn't parse remote URL")
 )
 
-func getWorkflowsForCurrentRepo(ghClient *ghapi.RESTClient, repo string) ([]ui.Workflow, error) {
+func getWorkflowsForCurrentRepo(ghClient *ghapi.RESTClient, repo string) ([]gh.Workflow, error) {
 	wd, err := gh.GetWorkflowDetails(ghClient, repo)
 	if err != nil {
 		return nil, err
 	}
 
-	workflows := make([]ui.Workflow, len(wd.Workflows))
+	workflows := make([]gh.Workflow, len(wd.Workflows))
 	for i, w := range wd.Workflows {
-		workflows[i] = ui.Workflow{
+		workflows[i] = gh.Workflow{
 			ID:   w.NodeID,
 			Repo: repo,
 			Name: w.Name,
