+++ internal/ui/assets/template.html
