+++ internal/ui/cmds.go
@@ -4,15 +4,16 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 	ghapi "github.com/cli/go-gh/v2/pkg/api"
 	ghgql "github.com/cli/shurcooL-graphql"
+	"github.com/dhth/act3/internal/gh"
 )
 
-func getWorkflowRuns(ghClient *ghapi.GraphQLClient, workflow Workflow) tea.Cmd {
+func getWorkflowRuns(ghClient *ghapi.GraphQLClient, workflow gh.Workflow) tea.Cmd {
 	return func() tea.Msg {
 		variables := map[string]interface{}{
 			"numWorkflowRuns": ghgql.Int(3),
 			"workflowId":      ghgql.ID(workflow.ID),
 		}
-		var query QueryResult
+		var query gh.QueryResult
 		err := ghClient.Query("GetWorkflows", &query, variables)
 
 		return WorkflowRunsFetchedMsg{workflow, query, err}
