+++ internal/ui/config.go
@@ -1,10 +1,13 @@
 package ui
 
-import ghapi "github.com/cli/go-gh/v2/pkg/api"
+import (
+	ghapi "github.com/cli/go-gh/v2/pkg/api"
+	"github.com/dhth/act3/internal/gh"
+)
 
 type Config struct {
 	GHClient     *ghapi.GraphQLClient
-	Workflows    []Workflow
+	Workflows    []gh.Workflow
 	CurrentRepo  *string
 	Fmt          OutputFmt
 	HTMLTemplate string
