--- ui/utils.go
@@ -1,25 +1,0 @@
-package ui
-
-import (
-	"strings"
-)
-
-func RightPadTrim(s string, length int) string {
-	if len(s) >= length {
-		if length > 3 {
-			return s[:length-3] + "..."
-		}
-		return s[:length]
-	}
-	return s + strings.Repeat(" ", length-len(s))
-}
-
-func Trim(s string, length int) string {
-	if len(s) >= length {
-		if length > 3 {
-			return s[:length-3] + "..."
-		}
-		return s[:length]
-	}
-	return s
-}
